+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added errors logging in the Crawler's `errorHandler` handler.
+
+### Fixed
+- Fixed the value of the field "finishedAt" in the result returned from Crawler.
 
 ## v0.5.1 - 2023-09-20
 ### Fixed